+++ CHANGELOG.md
@@ -2,3 +2,4 @@
  * Make macro-generated types (*RPCClient, *RPCServer) public
  * Split Transport into ClientTransport and ServerTransport
  * Pass state to rx_response
+ * Allow retrieving the underlying channel from built-in transports